+++ CHANGELOG.md
@@ -6,6 +6,9 @@ NOWrss 的所有改變將在此 CHANGELOG 文件中記錄。
 
 ## [Unreleased]
 
+### Change
+- yahoo 內文圖改用 imgapi 測試
+
 ## 1.1.6 - 2017-08-31
 ### Change
 - yahoo 文章內首圖改回之前的格式 @Webber